refactor: makes ProjectChecker an abstract class to mock it on tests
