+++ api/apps/api/src/modules/scenarios/scenarios.module.ts
@@ -41,12 +41,13 @@ import { SpecificationModule } from './specification';
 import { ScenarioFeaturesGapDataSerializer } from './dto/scenario-feature-gap-data.serializer';
 import { ScenarioFeaturesOutputGapDataSerializer } from './dto/scenario-feature-output-gap-data.serializer';
 import { CostRangeService } from './cost-range-service';
-import { ProjectChecker } from './project-checker.service';
+import { ProjectCheckerReal } from './project-checker.service-real';
 import { ProtectedAreaModule } from './protected-area';
 import { ProtectedAreasCrudModule } from '@marxan-api/modules/protected-areas/protected-areas-crud.module';
 import { PlanningAreasModule } from '@marxan-api/modules/planning-areas';
 import { BlmValuesModule } from '@marxan-api/modules/blm';
 import { BlmCalibrationModule } from './blm-calibration/blm-calibration.module';
+import { ProjectChecker } from '@marxan-api/modules/scenarios/project-checker.service';
 
 @Module({
   imports: [
@@ -82,7 +83,10 @@ import { BlmCalibrationModule } from './blm-calibration/blm-calibration.module';
     BlmCalibrationModule,
   ],
   providers: [
-    ProjectChecker,
+    {
+      useClass: ProjectCheckerReal,
+      provide: ProjectChecker,
+    },
     ScenariosService,
     ScenariosCrudService,
     ProxyService,
